third_party/WebKit/LayoutTests/editing/deleting/merge-paragraph-from-listing.html

<!DOCTYPE html>
<html> 
<body>
<p id="description">This tests deleting line break before listing.
WebKit should not cancel styles added by listing by those of document's default style.</p>
<div id="test" contenteditable>
hello
<listing>world</listing>
</div>
</div>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
<script>

function editingTest() {
    moveSelectionForwardByLineCommand();
    deleteCommand();

    Markup.description(document.getElementById('description').textContent);
    Markup.dump('test');
}

runDumpAsTextEditingTest(true);
</script>
</body>
</html>
