third_party/blink/web_tests/platform/generic/editing/deleting/merge-paragraph-from-listing-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
This tests deleting line break before listing.
WebKit should not cancel styles added by listing by those of document's default style.
| "
hello<#selection-caret>world"
